fix(clips): address start marker review
