+++ crates/vibez-engine/src/render.rs
@@ -249,7 +249,7 @@ fn render_offline_inner(
                     audio: Arc::clone(audio),
                     position: clip.position,
                     source_offset: clip.source_offset,
-                    start_marker: clip.start_marker.unwrap_or(clip.source_offset),
+                    start_marker: clip.resolved_start_marker(audio.num_frames() as u64),
                     duration: clip.duration,
                     loop_enabled: clip.loop_enabled,
                     loop_start: clip.loop_start,
@@ -273,7 +273,7 @@ fn render_offline_inner(
                 nc.position_beats,
                 nc.duration_beats,
                 nc.notes.clone(),
-                nc.start_marker_beats.unwrap_or(nc.loop_start_beats),
+                nc.resolved_start_marker_beats(),
                 nc.loop_enabled,
                 nc.loop_start_beats,
                 nc.loop_end_beats,
