+++ crates/vibez-ui/src/app/project_sections.rs
@@ -100,11 +100,6 @@ pub(super) fn timeline_without_audio(info: &TimelineInfo) -> ArrangementTimeline
         timeline.ensure(automation.track_id).automation = automation.lanes.clone();
     }
     for clip in &info.note_clips {
-        let marker_end = if clip.loop_enabled {
-            clip.duration_beats.min(clip.loop_end_beats)
-        } else {
-            clip.duration_beats
-        };
         timeline.ensure(clip.track_id).note_clips.push(UiNoteClip {
             id: clip.id,
             name: clip.name.clone(),
@@ -112,10 +107,7 @@ pub(super) fn timeline_without_audio(info: &TimelineInfo) -> ArrangementTimeline
             duration_beats: clip.duration_beats,
             notes: clip.notes.clone(),
             selected_notes: HashSet::new(),
-            start_marker_beats: clip
-                .start_marker_beats
-                .unwrap_or(clip.loop_start_beats)
-                .clamp(0.0, (marker_end - 0.01).max(0.0)),
+            start_marker_beats: clip.resolved_start_marker_beats(),
             loop_enabled: clip.loop_enabled,
             loop_start_beats: clip.loop_start_beats,
             loop_end_beats: clip.loop_end_beats,
@@ -142,18 +134,9 @@ pub(super) fn install_loaded_clip(
     loaded: crate::message::LoadedClipData,
 ) {
     let source_offset = loaded.info.source_offset;
-    let source_end = source_offset
-        .saturating_add(loaded.info.duration)
-        .min(loaded.audio.num_frames() as u64);
-    let marker_end = if loaded.info.loop_enabled {
-        source_end.min(loaded.info.loop_end)
-    } else {
-        source_end
-    };
-    let start_marker = loaded.info.start_marker.unwrap_or(source_offset).clamp(
-        source_offset,
-        marker_end.saturating_sub(1).max(source_offset),
-    );
+    let start_marker = loaded
+        .info
+        .resolved_start_marker(loaded.audio.num_frames() as u64);
     timeline.ensure(loaded.info.track_id).clips.push(UiClip {
         id: loaded.info.id,
         name: loaded.info.name,
