+++ crates/vibez-ui/src/app/views_perform_sections.rs
@@ -515,6 +515,7 @@ impl App {
                                     note.pitch,
                                     (note.start_beat - preview.position_beats).max(0.0),
                                     note.duration_beats,
+                                    false,
                                 )
                             })
                             .collect(),
